+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Routes, Route } from 'react-router';
 import { ThemeProvider } from 'styled-components';
-import { useAppSelector } from './store/hooks';
+import { useAppSelector } from './store';
 import { Layout, MainFC, MainCC, SettingsFC, SettingsCC, NotFound } from './pages';
 import { paths, lightTheme, darkTheme } from './constants';
 import { GlobalStyles } from './styled';
