+++ src/components/History/HistoryCC/interface.ts
@@ -1,0 +1,5 @@
+import { ICalculatorState } from '../../../store/slices/calculatorSlice';
+
+export interface IHistoryCC {
+  calculator: ICalculatorState;
+}
