+++ src/components/Keypad/KeypadCC/interface.ts
@@ -1,0 +1,9 @@
+import { ICalculatorState } from '../../../store/slices/calculatorSlice';
+
+export interface IKeypadCC {
+  calculator: ICalculatorState;
+  setInputValue: any;
+  deleteInputValue: any;
+  addToHistory: any;
+  equalInputValue: any;
+}
