+++ src/index.tsx
@@ -1,8 +1,8 @@
-import { createRoot } from 'react-dom/client';
 import { StrictMode } from 'react';
 import { HashRouter } from 'react-router-dom';
+import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { store } from './store/store';
+import { store } from './store';
 import { App } from './App';
 
 const container = document.getElementById('root') as HTMLDivElement;
