+++ src/store/hooks/index.ts
@@ -1,0 +1,4 @@
+import { useAppSelector, useAppDispatch } from './hooks';
+import { useBodyOverflow } from './useBodyOverflow';
+
+export { useAppSelector, useAppDispatch, useBodyOverflow };
