+++ src/store/index.ts
@@ -1,0 +1,35 @@
+import { store } from './store';
+import { useAppSelector, useAppDispatch, useBodyOverflow } from './hooks';
+import {
+  setInputValue,
+  addToHistory,
+  deleteInputValue,
+  equalInputValue,
+  resetAllCalculator,
+  resetInputValue,
+  openMobileMenu,
+  closeMobileMenu,
+  toggleMobileMenu,
+  switchLightTheme,
+  switchDarkTheme,
+  switchTheme,
+} from './slices';
+
+export {
+  store,
+  useAppSelector,
+  useAppDispatch,
+  useBodyOverflow,
+  setInputValue,
+  addToHistory,
+  deleteInputValue,
+  equalInputValue,
+  resetAllCalculator,
+  resetInputValue,
+  openMobileMenu,
+  closeMobileMenu,
+  toggleMobileMenu,
+  switchLightTheme,
+  switchDarkTheme,
+  switchTheme,
+};
